Density Learner

This Learner specializes Learner for density estimation problems:

task_type is set to "dens"

Creates Predictions of class PredictionDens.

Possible values for predict_types are:

"pdf": Evaluates estimated probability density function for each value in the test set.

"cdf": Evaluates estimated cumulative distribution function for each value in the test set.

See also

Super class

mlr3::Learner -> LearnerDens

Methods

Public methods

LearnerDens$new()

LearnerDens$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerDens$new(
  id,
  param_set = ps(),
  predict_types = "cdf",
  feature_types = character(),
  properties = character(),
  packages = character(),
  label = NA_character_,
  man = NA_character_
)

Arguments

id

(character(1))
Identifier for the new instance.

param_set

(paradox::ParamSet)
Set of hyperparameters.

predict_types

(character())
Supported predict types. Must be a subset of mlr_reflections$learner_predict_types.

feature_types

(character())
Feature types the learner operates on. Must be a subset of mlr_reflections$task_feature_types.

properties

(character())
Set of properties of the Learner (see initialization method $new(). Must be a subset of mlr_reflections$learner_properties.

packages

(character())
Set of required packages. A warning is signaled by the constructor if at least one of the packages is not installed, but loaded (not attached) later on-demand via requireNamespace().

label

(character(1))
Label for the new instance.

man

(character(1))
String in the format [pkg]::[topic] pointing to a manual page for this object. The referenced help package can be opened via method $help().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerDens$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
# get all density learners from mlr_learners:
lrns = mlr_learners$mget(mlr_learners$keys("^dens"))
names(lrns)
#>  [1] "dens.hist"      "dens.kde"       "dens.kde_ks"    "dens.locfit"   
#>  [5] "dens.logspline" "dens.mixed"     "dens.nonpar"    "dens.pen"      
#>  [9] "dens.plug"      "dens.spline"   

# get a specific learner from mlr_learners:
mlr_learners$get("dens.hist")
#> 
#> ── <LearnerDensHistogram> (dens.hist): Histogram Density Estimator ─────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3proba, and distr6
#> • Predict Types: [pdf], cdf, and distr
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties:
#> • Other settings: use_weights = 'error'
lrn("dens.hist")
#> 
#> ── <LearnerDensHistogram> (dens.hist): Histogram Density Estimator ─────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3proba, and distr6
#> • Predict Types: [pdf], cdf, and distr
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties:
#> • Other settings: use_weights = 'error'