Generalized Boosting Regression Modeling Survival Learner

Calls gbm::gbm().

lp is predicted by gbm::predict.gbm()

crank is identical to lp

Parameter distribution is set to coxph as this is the only distribution implemented in gbm::gbm() for survival analysis; parameter keep.data is set to FALSE for efficiency.

Format

R6::R6Class() inheriting from LearnerSurv.

Construction

LearnerSurvGBM$new()
mlr_learners$get("surv.gbm")
lrn("surv.gbm")

Meta Information

Type: "surv"

Predict Types: crank, lp

Feature Types: integer, numeric, factor, ordered

Packages: gbm

References

Freund Y, Schapire RE (1997). “A Decision-Theoretic Generalization of On-Line Learning and an Application to Boosting.” Journal of Computer and System Sciences, 55(1), 119--139. doi: 10.1006/jcss.1997.1504 , https://doi.org/10.1006/jcss.1997.1504.

Ridgeway G (1999). “The state of boosting.” Computing Science and Statistics, 172--181.

Friedman J, Hastie T, Tibshirani R (2000). “Additive logistic regression: a statistical view of boosting (With discussion and a rejoinder by the authors).” The Annals of Statistics, 28(2), 337--407. doi: 10.1214/aos/1016218223 , https://doi.org/10.1214/aos/1016218223.

Friedman JH (2001). “Greedy Function Approximation: A Gradient Boosting Machine.” The Annals of Statistics, 29(5), 1189--1232. http://www.jstor.org/stable/2699986.

Friedman JH (2001). “Greedy Function Approximation: A Gradient Boosting Machine.” The Annals of Statistics, 29(5), 1189--1232. http://www.jstor.org/stable/2699986.

Friedman JH (2002). “Stochastic gradient boosting.” Computational Statistics & Data Analysis, 38(4), 367--378. doi: 10.1016/s0167-9473(01)00065-2 , https://doi.org/10.1016/s0167-9473(01)00065-2.

Kriegler B (2007). Cost-Sensitive Stochastic Gradient Boosting Within a Quantitative Regression Framework. Ph.D. thesis, University of California at Los Angeles. https://dl.acm.org/citation.cfm?id=1354603.

Burges CJ (2010). “From RankNet to LambdaRank to LambdaMART: An Overview.” Technical Report MSR-TR-2010-82, Microsoft Research. https://www.microsoft.com/en-us/research/publication/from-ranknet-to-lambdarank-to-lambdamart-an-overview/.

See also