Fully Parametric Survival Learner

Calls survival::survreg().

lp is predicted by using an internally defined predict method, see details

distr is predicted by using an internally defined predict method, see details

crank is identical to lp.

Currently six parameterisations can be assumed for the baseline, see survival::survreg(). These are internally re-parameterised and defined as distr6 objects, we plan on implementing more custom distributions in the future.

Format

R6::R6Class() inheriting from LearnerSurv.

Details

The internal predict method is implemented in mlr3proba, which is more efficient for composition to distributions than survival::predict.survreg().

lp is predicted using the formula \(lp = X\beta\) where \(X\) are the variables in the test data set and \(\beta\) are the fitted coefficients.

The distribution distr is composed using the lp and specifying a model form in the type hyper-parameter. These are as follows, with respective hazard functions,

Proportional Hazards (ph) $$h(t) = h_0(t)exp(lp)$$

Accelerated Failure Time (aft) $$h(t) = exp(-lp)h_0(\frac{t}{exp(lp)})$$

Proportional Odds (po) $$\frac{h(t)}{h_0(t)} = {1 + (exp(lp) - 1)S_0(t)}^{-1}$$

Where \(h_0\) and \(S_0\) are the estimated baseline hazard/survival respectively (in this case with a given parametric form), and \(lp\) is the predicted linear predictor lp.

Construction

LearnerSurvParametric$new()
mlr_learners$get("surv.parametric")
lrn("surv.parametric")

Meta Information

Type: "surv"

Predict Types: distr, lp, crank

Feature Types: logical, integer, numeric, factor

Packages: survival distr6

References

Kalbfleisch, J. D., Prentice, R. L. (2002). The Statistical Analysis of Failure Time Data. John Wiley & Sons. doi: 10.1002/9781118032985 .

See also