Regression, Ranking and Hybrid Support Vector Machines Survival Learner

Calls survivalsvm::survivalsvm().

crank is predicted by survivalsvm::predict.survivalsvm

Four possible SVMs can be implemented, dependent on the type parameter. These correspond to predicting the survival time via regression (regression), predicting the PI by imposing a ranking constraint (vanbelle1, vanbelle2), a hybrid of the two (hybrid). Whichever type is chosen determines how the crank predict type is calculated, but in any case all can be considered a valid continuous ranking.

To be in line with the Van Belle papers and to prevent the learner crashing without user-set parameters, the default diff.meth is set to diffmeth3 with gamma.mu equal to 0.1.

Format

R6::R6Class() inheriting from LearnerSurv.

Construction

LearnerSurvSVM$new()
mlr_learners$get("surv.svm")
lrn("surv.svm")

Meta Information

Type: "surv"

Predict Types: crank

Feature Types: integer, numeric

Packages: survivalsvm

References

Van Belle, V., Pelcmans, K., Van Huffel S. and Suykens J. A.K. (2011a). Improved performance on high-dimensional survival data by application of Survival-SVM. Bioinformatics, 27, 87-94.

Van Belle, V., Pelcmans, K., Van Huffel S. and Suykens J. A.K. (2011b). Support vector methods for survival analysis: a comparison between ranking and regression approaches. Artificial Intelligence in Medicine, 53, 107-118.

See also

Examples

library(mlr3)
task = tgen("simsurv")$generate(200)
learner = lrn("surv.svm")
resampling = rsmp("cv", folds = 3)
resample(task, learner, resampling)

#> <ResampleResult> of 3 iterations
#> * Task: simsurv
#> * Learner: surv.svm
#> * Warnings: 0 in 0 iterations
#> * Errors: 0 in 0 iterations