Standard Error of Integrated Log loss Survival Measure

Calculates the standard error of MeasureSurvIntLogloss.

If integrated == FALSE then the standard error of the loss, L, is approximated via, $$se(L) = sd(L)/\sqrt{N}$$ where \(N\) are the number of observations in the test set, and \(sd\) is the standard deviation.

If integrated == TRUE then correlations between time-points need to be taken into account, therefore $$se(L) = \sqrt{\frac{\sum_{i = 1}^M\sum_{j=1}^M \Sigma_{i,j}}{NT^2}}$$ where \(\Sigma_{i, j}\) is the sample covariance matrix over \(M\) distinct time-points.

Format

R6::R6Class() inheriting from MeasureSurvIntegrated/MeasureSurv.

Construction

MeasureSurvIntLoglossSE$new(integrated = TRUE, times, eps = 1e-15)
mlr_measures$get("surv.intloglossSE")
msr("surv.intloglossSE")

integrated :: logical(1)
If TRUE (default), returns the integrated score; otherwise, not integrated.

times :: vector()
If integrate == TRUE then a vector of time-points over which to integrate the score. If integrate == FALSE then a single time point at which to return the score.

eps :: numeric(1)
Very small number to set zero-valued predicted probabilities to, in order to prevent errors in log(0) calculation.

Meta Information

Type: "surv"

Range: \([0, \infty)\)

Minimize: TRUE

Required prediction: distr

Fields

See MeasureSurv, as well as all variables passed to the constructor.

As well as

eps :: numeric(1)
Very small number to set zero-valued predicted probabilities to, in order to prevent errors in log(0) calculation.

References

Graf E, Schmoor C, Sauerbrei W, Schumacher M (1999). “Assessment and comparison of prognostic classification schemes for survival data.” Statistics in Medicine, 18(17-18), 2529--2545. doi: 10.1002/(sici)1097-0258(19990915/30)18:17/18<2529::aid-sim274>3.0.co;2-5 .

See also