Abstract Class for Integrated Measures

This is an abstract class that should not be constructed directly.

Super classes

mlr3::Measure -> mlr3proba::MeasureSurv -> MeasureSurvIntegrated

Active bindings

eps

(numeric(1))
Very small number used to prevent log(0) and 1/0 error.

integrated

(logical(1)) Returns if the measure should be integrated or not. Settable.

times

(numeric()) Returns the times at which the measure should be evaluated at, or integrated over. Settable.

method

(integer(1)) Returns which method is used for approximating integration. Settable.

proper

(logical(1))
If TRUE then weights scores by the censoring distribution at the observed event time, which results in a strictly proper scoring rule if censoring and survival time distributions are independent and a sufficiently large dataset is used to weight the measure. If FALSE then weights scores by the Graf method which is the more common usage but the loss is not proper.

Methods

Public methods

MeasureSurvIntegrated$new()

MeasureSurvIntegrated$clone()

Inherited methods

mlr3::Measure$aggregate()

mlr3::Measure$format()

mlr3::Measure$help()

mlr3::Measure$score()

mlr3proba::MeasureSurv$print()

Method new()

This is an abstract class that should not be constructed directly.

Usage

MeasureSurvIntegrated$new(
  integrated = TRUE,
  times,
  method = 2,
  proper = FALSE,
  eps = 0.000000000000001,
  id,
  range,
  minimize,
  packages,
  predict_type,
  properties = character(),
  man = NA_character_,
  se = FALSE
)

Arguments

integrated

(logical(1))
If TRUE (default), returns the integrated score; otherwise, not integrated.

times

(numeric())
If integrate == TRUE then a vector of time-points over which to integrate the score. If integrate == FALSE then a single time point at which to return the score.

method

(integer(1))
If integrate == TRUE selects the integration weighting method. method == 1 corresponds to weighting each time-point equally and taking the mean score over discrete time-points. method == 2 corresponds to calculating a mean weighted by the difference between time-points. method == 2 is default to be in line with other packages.

proper

(logical(1))
If TRUE then weights scores by the censoring distribution at the observed event time, which results in a strictly proper scoring rule if censoring and survival time distributions are independent and a sufficiently large dataset is used to weight the measure. If FALSE then weights scores by the Graf method which is the more common usage but the loss is not proper. In v0.5.0, the default will be changed to TRUE.

eps

(numeric(1))
Very small number to set zero-valued predicted probabilities to in order to prevent errors in log(0) and 1/0 calculation.

id

(character(1))
Identifier for the new instance.

range

(numeric(2))
Feasible range for this measure as c(lower_bound, upper_bound). Both bounds may be infinite.

minimize

(logical(1))
Set to TRUE if good predictions correspond to small values, and to FALSE if good predictions correspond to large values. If set to NA (default), tuning this measure is not possible.

packages

(character())
Set of required packages. A warning is signaled by the constructor if at least one of the packages is not installed, but loaded (not attached) later on-demand via requireNamespace().

predict_type

(character(1))
Required predict type of the Learner. Possible values are stored in mlr_reflections$learner_predict_types.

properties

(character())
Properties of the measure. Must be a subset of mlr_reflections$measure_properties. Supported by mlr3:

"requires_task" (requires the complete Task),

"requires_learner" (requires the trained Learner),

"requires_train_set" (requires the training indices from the Resampling), and

"na_score" (the measure is expected to occasionally return NA or NaN).

man

(character(1))
String in the format [pkg]::[topic] pointing to a manual page for this object. The referenced help package can be opened via method $help().

se

(logical(1))
If TRUE returns the standard error of the measure.

Method clone()

The objects of this class are cloneable with this method.

Usage

MeasureSurvIntegrated$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.