Song and Zhou's AUC Survival Measure

Calls survAUC::AUC.sh().

Assumes Cox PH model specification.

Details

All measures implemented from survAUC should be used with care, we are aware of problems in implementation that sometimes cause fatal errors in R. In future updates these measures will all be re-written and implemented directly in mlr3proba.

Dictionary

This Measure can be instantiated via the dictionary mlr_measures or with the associated sugar function msr():

MeasureSurvSongAUC$new()
mlr_measures$get("surv.songAUC")
msr("surv.songAUC")

Meta Information

Type: "surv"

Range: \([0, 1]\)

Minimize: FALSE

Required prediction: lp

References

Song X, Zhou X (2008). “A semiparametric approach for the covariate specific ROC curve with survival outcome.” Statistica Sinica, 18(3), 947--65. http://www.jstor.org/stable/24308524.

See also

Super classes

mlr3::Measure -> mlr3proba::MeasureSurv -> mlr3proba::MeasureSurvIntegrated -> mlr3proba::MeasureSurvAUC -> MeasureSurvSongAUC

Active bindings

type

(character(1))
Type of measure, one of: 'cumulative', 'incident'.

Methods

Public methods

MeasureSurvSongAUC$new()

MeasureSurvSongAUC$clone()

Inherited methods

mlr3::Measure$aggregate()

mlr3::Measure$format()

mlr3::Measure$help()

mlr3::Measure$print()

mlr3::Measure$score()

Method new()

Creates a new instance of this R6 class.

Usage

MeasureSurvSongAUC$new(
  integrated = TRUE,
  times,
  type = c("incident", "cumulative")
)

Arguments

integrated

(logical(1))
If TRUE (default), returns the integrated score; otherwise, not integrated.

times

(numeric())
If integrate == TRUE then a vector of time-points over which to integrate the score. If integrate == FALSE then a single time point at which to return the score.

type

(character(1))
Determines the type of score, one of: 'cumulative', 'incident'.

Method clone()

The objects of this class are cloneable with this method.

Usage

MeasureSurvSongAUC$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.