PipeOpPredTransformer

Parent class for PipeOps that transform Prediction objects to different types.

Input and Output Channels

PipeOpPredTransformer has one input and output channel named "input" and "output". In training and testing these expect and produce mlr3::Prediction objects with the type depending on the transformers.

State

The $state is a named list with the $state elements

inpredtypes: Predict types in the input prediction object during training.

outpredtypes : Predict types in the input prediction object during prediction, checked against inpredtypes.

Internals

Classes inheriting from PipeOpPredTransformer transform Prediction objects from one class (e.g. regr, classif) to another.

See also

Super classes

mlr3pipelines::PipeOp -> mlr3proba::PipeOpTransformer -> PipeOpPredTransformer

Methods

Public methods

PipeOpPredTransformer$new()

PipeOpPredTransformer$clone()

Inherited methods

mlr3pipelines::PipeOp$predict()

mlr3pipelines::PipeOp$print()

mlr3pipelines::PipeOp$train()

Method new()

Creates a new instance of this R6 class.

Usage

PipeOpPredTransformer$new(
  id,
  param_set = ParamSet$new(),
  param_vals = list(),
  packages = character(0),
  input = data.table(),
  output = data.table()
)

Arguments

id

(character(1))
Identifier of the resulting object.

param_set

(paradox::ParamSet)
Set of hyperparameters.

param_vals

(list())
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction.

packages

(character())
Set of required packages. A warning is signaled by the constructor if at least one of the packages is not installed, but loaded (not attached) later on-demand via requireNamespace().

input

data.table::data.table
data.table with columns name (character), train (character), predict (character). Sets the $input slot, see PipeOp.

output

data.table::data.table
data.table with columns name (character), train (character), predict (character). Sets the $output slot, see PipeOp.

Method clone()

The objects of this class are cloneable with this method.

Usage

PipeOpPredTransformer$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.