PipeOpTransformer

Parent class for PipeOps that transform Task and Prediction objects to different types.

Input and Output Channels

Determined by child classes.

State

The $state is left empty (list()).

Internals

The commonality of methods using PipeOpTransformer is that they take a Task or Prediction of one type (e.g. regr or classif) and transform it to another type.

See also

Super class

mlr3pipelines::PipeOp -> PipeOpTransformer

Methods

Public methods

PipeOpTransformer$new()

PipeOpTransformer$clone()

Inherited methods

mlr3pipelines::PipeOp$help()

mlr3pipelines::PipeOp$predict()

mlr3pipelines::PipeOp$print()

mlr3pipelines::PipeOp$train()

Method new()

Creates a new instance of this R6 class.

Usage

PipeOpTransformer$new(
  id,
  param_set = ps(),
  param_vals = list(),
  packages = character(),
  input = data.table(),
  output = data.table()
)

Arguments

id

(character(1))
Identifier of the resulting object.

param_set

(paradox::ParamSet)
Set of hyperparameters.

param_vals

(list())
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction.

packages

(character())
Set of required packages. A warning is signaled by the constructor if at least one of the packages is not installed, but loaded (not attached) later on-demand via requireNamespace().

input

data.table::data.table
data.table with columns name (character), train (character), predict (character). Sets the $input slot, see PipeOp.

output

data.table::data.table
data.table with columns name (character), train (character), predict (character). Sets the $output slot, see PipeOp.

Method clone()

The objects of this class are cloneable with this method.

Usage

PipeOpTransformer$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.