Compose a Crank Predict Type for Survival Learners

This is a wrapper around the PipeOpCrankCompositor pipe operation, which simplifies graph creation.

crankcompositor(
  learner,
  method = c("mean", "median", "mode"),
  param_vals = list()
)

Arguments

learner
LearnerSurv object for which a crank is composed (or over-written)

method
One of mean, mode, or median; abbreviations allowed. Used to determine how crank is estimated from the predicted distr. Default is mean.

param_vals
Additional parameters to pass to the learner.

Value

mlr3pipelines::GraphLearner

Details

For full details see PipeOpCrankCompositor.

Examples

library(mlr3)
library(mlr3pipelines)

task = tgen("simsurv")$generate(20)
ranger.crank = crankcompositor(learner = lrn("surv.coxph"),
                            method = "median")
resample(task, ranger.crank, rsmp("cv", folds = 2))$predictions()

#> [[1]]
#> <PredictionSurv> for 10 observations:
#>     row_id      time status    crank                distr        lp
#>          2 0.3707709   TRUE 5.000000 <VectorDistribution> -3.088300
#>          7 5.0000000  FALSE 5.000000 <VectorDistribution> -1.302434
#>          8 2.4830070   TRUE 4.023879 <VectorDistribution>  2.052353
#> ---                                                                
#>         17 1.9768585   TRUE 4.382142 <VectorDistribution>  1.509337
#>         18 2.4669239   TRUE 5.000000 <VectorDistribution> -2.515227
#>         20 5.0000000  FALSE 5.000000 <VectorDistribution> -1.806704
#> 
#> [[2]]
#> <PredictionSurv> for 10 observations:
#>     row_id     time status    crank                distr          lp
#>          1 5.000000  FALSE 2.483007 <VectorDistribution>  0.25443229
#>          3 5.000000  FALSE 5.000000 <VectorDistribution> -0.24281955
#>          4 5.000000  FALSE 5.000000 <VectorDistribution> -0.04420689
#> ---                                                                 
#>         13 4.023879   TRUE 5.000000 <VectorDistribution>  0.02793521
#>         16 5.000000  FALSE 2.466924 <VectorDistribution>  0.56127953
#>         19 4.382142   TRUE 5.000000 <VectorDistribution> -0.21990905
#>