Compose a Distr Predict Type for Survival Learners

This is a wrapper around the PipeOpDistrCompositor pipe operation, which simplifies graph creation.

distrcompositor(
  learner,
  estimator = c("kaplan", "nelson"),
  form = c("aft", "ph", "po"),
  overwrite = FALSE,
  param_vals = list()
)

Arguments

learner
LearnerSurv object for which a distr is composed (or over-written).

estimator
One of kaplan or nelson, corresponding to the Kaplan-Meier and Nelson-Aalen estimators respectively. Used to estimate the baseline survival distribution. Abbreviations allowed. Default is kaplan.

form
One of aft, ph, or po, corresponding to accelerated failure time, proportional hazards, and proportional odds respectively. Used to determine the form of the composed survival distribution. Default is aft.

overwrite
logical. If FALSE (default) then if the learner already has a distr, the compositor does nothing. If TRUE then the distr is overwritten by the compositor if already present, which may be required for changing the prediction distr from one model form to another.

param_vals
Additional parameters to pass to the learner.

Value

mlr3pipelines::GraphLearner

Details

For full details see PipeOpDistrCompositor.

Examples

library("mlr3")
library("mlr3pipelines")

task = tgen("simsurv")$generate(20)
cvglm.distr = distrcompositor(learner = lrn("surv.cvglmnet"),
                            estimator = "kaplan",
                            form = "aft")
resample(task, cvglm.distr,
        rsmp("cv", folds = 2))$predictions()

#> [[1]]
#> <PredictionSurv> for 10 observations:
#>     row_id     time status     crank                distr        lp
#>          1 5.000000  FALSE -3.562780 <VectorDistribution> -3.562780
#>          2 5.000000  FALSE -2.532081 <VectorDistribution> -2.532081
#>          4 5.000000  FALSE -3.411464 <VectorDistribution> -3.411464
#> ---                                                                
#>         12 5.000000  FALSE -2.507915 <VectorDistribution> -2.507915
#>         17 3.139458   TRUE -3.028849 <VectorDistribution> -3.028849
#>         19 1.645720   TRUE -2.544916 <VectorDistribution> -2.544916
#> 
#> [[2]]
#> <PredictionSurv> for 10 observations:
#>     row_id      time status         crank                distr            lp
#>          3 2.9711146   TRUE -5.238021e-06 <VectorDistribution> -5.238021e-06
#>          6 0.5168292   TRUE -5.275918e-06 <VectorDistribution> -5.275918e-06
#>          9 5.0000000  FALSE -5.319039e-06 <VectorDistribution> -5.319039e-06
#> ---                                                                         
#>         16 4.8813723   TRUE -5.288955e-06 <VectorDistribution> -5.288955e-06
#>         18 5.0000000  FALSE -5.460378e-06 <VectorDistribution> -5.460378e-06
#>         20 5.0000000  FALSE -6.029435e-06 <VectorDistribution> -6.029435e-06
#>