KS Kernel Density Estimator

Calls ks::kde() and the result is coerced to a distr6::Distribution.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

LearnerDensKDEks$new()
mlr_learners$get("dens.kde_ks")
lrn("dens.kde_ks")

Meta Information

Type: "dens"

Predict Types: pdf

Feature Types: integer, numeric

Properties: -

Packages: mlr3 mlr3proba ks distr6

References

Gramacki, Artur, Gramacki, Jarosław (2017). “FFT-based fast computation of multivariate kernel density estimators with unconstrained bandwidth matrices.” Journal of Computational and Graphical Statistics, 26(2), 459–462.

See also

Super classes

mlr3::Learner -> mlr3proba::LearnerDens -> LearnerDensKDEks

Methods

Public methods

LearnerDensKDEks$new()

LearnerDensKDEks$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerDensKDEks$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerDensKDEks$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("dens.kde_ks")
print(learner)
#> 
#> ── <LearnerDensKDEks> (dens.kde_ks): Kernel Density Estimator (ks package) ─────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3proba, ks, and distr6
#> • Predict Types: [pdf]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties:
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("faithful")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> ksKDE() 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> dens.logloss 
#>    0.9899349