Cox Proportional Hazards Survival Learner

Calls survival::coxph().

lp is predicted by survival::predict.coxph()

distr is predicted by survival::survfit.coxph()

crank is identical to lp

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

LearnerSurvCoxPH$new()
mlr_learners$get("surv.coxph")
lrn("surv.coxph")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”, “factor”

Required Packages: mlr3, mlr3proba, survival, distr6

Parameters

Id	Type	Default	Levels	Range

ties	character	efron	efron, breslow, exact	-

singular.ok	logical	TRUE	TRUE, FALSE	-

type	character	efron	efron, aalen, kalbfleisch-prentice	-

stype	integer	2		\([1, 2]\)

References

Cox DR (1972). “Regression Models and Life-Tables.” Journal of the Royal Statistical Society: Series B (Methodological), 34(2), 187–202. doi:10.1111/j.2517-6161.1972.tb00899.x .

See also

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvCoxPH

Methods

Public methods

LearnerSurvCoxPH$new()

LearnerSurvCoxPH$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvCoxPH$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvCoxPH$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.