Kaplan-Meier Estimator Survival Learner

Calls survival::survfit().

distr is predicted by estimating the survival function with survival::survfit()

crank is predicted as the sum of the cumulative hazard function (expected mortality) derived from the survival distribution, distr

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

LearnerSurvKaplan$new()
mlr_learners$get("surv.kaplan")
lrn("surv.kaplan")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”

Feature Types: “logical”, “integer”, “numeric”, “character”, “factor”, “ordered”

Required Packages: mlr3, mlr3proba, survival, distr6

Parameters

Empty ParamSet

References

Kaplan EL, Meier P (1958). “Nonparametric Estimation from Incomplete Observations.” Journal of the American Statistical Association, 53(282), 457–481. doi:10.1080/01621459.1958.10501452 .

See also

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvKaplan

Methods

Public methods

LearnerSurvKaplan$new()

LearnerSurvKaplan$importance()

LearnerSurvKaplan$selected_features()

LearnerSurvKaplan$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvKaplan$new()

Method importance()

All features have a score of 0 for this learner. This method exists solely for compatibility with the mlr3 ecosystem, as this learner is used as a fallback for other survival learners that require an importance() method.

Usage

LearnerSurvKaplan$importance()

Returns

Named numeric().

Method selected_features()

Selected features are always the empty set for this learner. This method is implemented only for compatibility with the mlr3 API, as this learner does not perform feature selection.

Usage

LearnerSurvKaplan$selected_features()

Returns

character(0).

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvKaplan$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.