Blanche's AUC Competing Risks Measure

Calculates the cause-specific ROC-AUC(t) at a specific time point, see Blanche et al. (2013). Can also return the mean AUC(t) over all competing causes.

Details

Calls riskRegression::Score() with:

metric = "auc"

cens.method = "ipcw"

cens.model = "km"

Note that the IPC weights (estimated via the Kaplan-Meier) are calculated using the test data.

Dictionary

This Measure can be instantiated via the dictionary mlr_measures or with the associated sugar function msr():

MeasureCompRisksAUC$new()
mlr_measures$get("cmprsk.auc")
msr("cmprsk.auc")

Parameters

Id	Type	Default	Range

cause	integer	1	\([1, \infty)\)

time_horizon	numeric	NULL	\([0, \infty)\)

Meta Information

Type: "cmprsk"

Range: \([0, 1]\)

Minimize: FALSE

Required prediction: cif

Parameter details

cause (numeric(1))
Integer number indicating which cause to use (Default: 1). If "mean", then the mean AUC(t) over all causes is returned.

time_horizon (numeric(1))
Single time point at which to return the score. If NULL, we issue a warning and the median time from the test set is used.

References

Blanche, Paul, Dartigues, Francois J, Jacqmin-Gadda, Helene (2013). “Estimating and comparing time-dependent areas under receiver operating characteristic curves for censored event times with competing risks.” Statistics in Medicine, 32(30), 5381–5397. ISSN 1097-0258, doi:10.1002/SIM.5958 , https://onlinelibrary.wiley.com/doi/10.1002/sim.5958.

Super classes

mlr3::Measure -> mlr3proba::MeasureCompRisks -> MeasureCompRisksAUC

Methods

Public methods

MeasureCompRisksAUC$new()

MeasureCompRisksAUC$clone()

Inherited methods

mlr3::Measure$aggregate()

mlr3::Measure$format()

mlr3::Measure$help()

mlr3::Measure$print()

mlr3::Measure$score()

Method new()

Creates a new instance of this R6 class.

Usage

MeasureCompRisksAUC$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

MeasureCompRisksAUC$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

t = tsk("pbc")
l = lrn("cmprsk.aalen")
p = l$train(t)$predict(t)

p$score(msr("cmprsk.auc", time_horizon = 42))
#> cmprsk.auc 
#>        0.5