Log Loss Density Measure

Calculates the cross-entropy, or logarithmic (log), loss.

Details

The Log Loss, in the context of probabilistic predictions, is defined as the negative log probability density function, \(f\), evaluated at the observed value, \(y\), $$L(f, y) = -\log(f(y))$$

Dictionary

This Measure can be instantiated via the dictionary mlr_measures or with the associated sugar function msr():

MeasureDensLogloss$new()
mlr_measures$get("dens.logloss")
msr("dens.logloss")

Parameters

Id	Type	Default	Range

eps	numeric	1e-15	\([0, 1]\)

Meta Information

Type: "density"

Range: \([0, \infty)\)

Minimize: TRUE

Required prediction: pdf

Parameter details

eps (numeric(1))
Very small number to substitute zero values in order to prevent errors in e.g. log(0) and/or division-by-zero calculations. Default value is 1e-15.

Super classes

mlr3::Measure -> mlr3proba::MeasureDens -> MeasureDensLogloss

Methods

Public methods

MeasureDensLogloss$new()

MeasureDensLogloss$clone()

Inherited methods

mlr3::Measure$aggregate()

mlr3::Measure$format()

mlr3::Measure$help()

mlr3::Measure$print()

mlr3::Measure$score()

Method new()

Creates a new instance of this R6 class.

Usage

MeasureDensLogloss$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

MeasureDensLogloss$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.