Concordance Statistics Survival Measure

Calculates weighted concordance statistics, which, depending on the chosen weighting method and tied times solution, are equivalent to several proposed methods.

For the Kaplan-Meier estimate of the training survival distribution, S, and the Kaplan-Meier estimate of the training censoring distribution, G:

weight_meth:

"I" = No weighting. (Harrell)

"GH" = Gonen and Heller's Concordance Index

"G" = Weights concordance by G^-1.

"G2" = Weights concordance by G^-2. (Uno et al.)

"SG" = Weights concordance by S/G (Shemper et al.)

"S" = Weights concordance by S (Peto and Peto)

The last three require training data. "GH" is only applicable to LearnerSurvCoxPH.

@details The implementation is slightly different from survival::concordance. Firstly this implementation is faster, and secondly the weights are computed on the training dataset whereas in survival::concordance the weights are computed on the same testing data.

Dictionary

This Measure can be instantiated via the dictionary mlr_measures or with the associated sugar function msr():

MeasureSurvCindex$new()
mlr_measures$get("surv.cindex")
msr("surv.cindex")

Meta Information

Type: "surv"

Range: \([0, 1]\)

Minimize: FALSE

Required prediction: crank

References

Peto, Richard, Peto, Julian (1972). “Asymptotically efficient rank invariant test procedures.” Journal of the Royal Statistical Society: Series A (General), 135(2), 185--198.

Harrell, E F, Califf, M R, Pryor, B D, Lee, L K, Rosati, A R (1982). “Evaluating the yield of medical tests.” Jama, 247(18), 2543--2546.

Gönen M, Heller G (2005). “Concordance probability and discriminatory power in proportional hazards regression.” Biometrika, 92(4), 965--970. doi: 10.1093/biomet/92.4.965 .

Schemper, Michael, Wakounig, Samo, Heinze, Georg (2009). “The estimation of average hazard ratios by weighted Cox regression.” Statistics in Medicine, 28(19), 2473--2489. doi: 10.1002/sim.3623 .

Uno H, Cai T, Pencina MJ, D'Agostino RB, Wei LJ (2011). “On the C-statistics for evaluating overall adequacy of risk prediction procedures with censored survival data.” Statistics in Medicine, n/a--n/a. doi: 10.1002/sim.4154 .

See also

Super classes

mlr3::Measure -> mlr3proba::MeasureSurv -> MeasureSurvCindex

Active bindings

cutoff

(numeric(1)) Cut-off time to evaluate concordance up to.

weight_meth

(numeric(1)) Method for weighting concordance.

tiex

(numeric(1)) Cut-off time to evaluate concordance up to.

Methods

Public methods

MeasureSurvCindex$new()

MeasureSurvCindex$clone()

Inherited methods

mlr3::Measure$aggregate()

mlr3::Measure$format()

mlr3::Measure$help()

mlr3::Measure$score()

mlr3proba::MeasureSurv$print()

Method new()

This is an abstract class that should not be constructed directly.

Usage

MeasureSurvCindex$new(
  cutoff = NULL,
  weight_meth = c("I", "G", "G2", "SG", "S", "GH"),
  tiex = 0.5
)

Arguments

cutoff

(numeric(1))
Cut-off time to evaluate concordance up to.

weight_meth

(character(1))
Method for weighting concordance. Default "I" is Harrell's C. See details.

tiex

(numeric(1))
Weighting applied to tied rankings, default is to give them half weighting.

Method clone()

The objects of this class are cloneable with this method.

Usage

MeasureSurvCindex$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.