Song and Zhou's TNR Survival Measure

Calls survAUC::spec.sh().

Assumes Cox PH model specification.

times and lp_thresh are arbitrarily set to 0 to prevent crashing, these should be further specified.

Details

All measures implemented from survAUC should be used with care, we are aware of problems in implementation that sometimes cause fatal errors in R. In future updates some of these measures may be re-written and implemented directly in mlr3proba.

Dictionary

This Measure can be instantiated via the dictionary mlr_measures or with the associated sugar function msr():

MeasureSurvSongTNR$new()
mlr_measures$get("surv.song_tnr")
msr("surv.song_tnr")

Parameters

Id	Type	Default	Range

times	numeric	-	\([0, \infty)\)

lp_thresh	numeric	0	\((-\infty, \infty)\)

Meta Information

Type: "surv"

Range: \([0, 1]\)

Minimize: FALSE

Required prediction: lp

Parameter details

times (numeric())
If integrated == TRUE then a vector of time-points over which to integrate the score. If integrated == FALSE then a single time point at which to return the score.

lp_thresh (numeric(1))
Determines the cutoff threshold of the linear predictor in the calculation of the TPR/TNR scores.

References

Song, Xiao, Zhou, Xiao-Hua (2008). “A semiparametric approach for the covariate specific ROC curve with survival outcome.” Statistica Sinica, 18(3), 947–65. https://www.jstor.org/stable/24308524.

See also

Super classes

mlr3::Measure -> mlr3proba::MeasureSurv -> mlr3proba::MeasureSurvAUC -> MeasureSurvSongTNR

Methods

Public methods

MeasureSurvSongTNR$new()

MeasureSurvSongTNR$clone()

Inherited methods

mlr3::Measure$aggregate()

mlr3::Measure$format()

mlr3::Measure$help()

mlr3::Measure$print()

mlr3::Measure$score()

Method new()

Creates a new instance of this R6 class.

Usage

MeasureSurvSongTNR$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

MeasureSurvSongTNR$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.