Uno's TPR Survival Measure

Calls survAUC::sens.uno().

Assumes random censoring.

times and lp_thresh are arbitrarily set to 0 to prevent crashing, these should be further specified.

Details

All measures implemented from survAUC should be used with care, we are aware of problems in implementation that sometimes cause fatal errors in R. In future updates these measures will all be re-written and implemented directly in mlr3proba.

Dictionary

This Measure can be instantiated via the dictionary mlr_measures or with the associated sugar function msr():

MeasureSurvUnoTPR$new()
mlr_measures$get("surv.uno_tpr")
msr("surv.uno_tpr")

Meta Information

Type: "surv"

Range: \([0, 1]\)

Minimize: FALSE

Required prediction: lp

References

Uno H, Cai T, Tian L, Wei LJ (2007). “Evaluating Prediction Rules fort-Year Survivors With Censored Regression Models.” Journal of the American Statistical Association, 102(478), 527--537. doi: 10.1198/016214507000000149 .

See also

Super classes

mlr3::Measure -> mlr3proba::MeasureSurv -> mlr3proba::MeasureSurvIntegrated -> mlr3proba::MeasureSurvAUC -> MeasureSurvUnoTPR

Active bindings

lp_thresh

numeric(1)
Threshold for linear predictor when calculating TPR/TNR.

Methods

Public methods

MeasureSurvUnoTPR$new()

MeasureSurvUnoTPR$clone()

Inherited methods

mlr3::Measure$aggregate()

mlr3::Measure$format()

mlr3::Measure$help()

mlr3::Measure$score()

mlr3proba::MeasureSurv$print()

Method new()

Creates a new instance of this R6 class.

Usage

MeasureSurvUnoTPR$new(times = 0, lp_thresh = 0)

Arguments

times

(numeric())
If integrate == TRUE then a vector of time-points over which to integrate the score. If integrate == FALSE then a single time point at which to return the score.

lp_thresh

numeric(1)
Determines where to threshold the linear predictor for calculating the TPR/TNR.

Method clone()

The objects of this class are cloneable with this method.

Usage

MeasureSurvUnoTPR$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.