Xu and O'Quigley's R2 Survival Measure

Calls survAUC::XO().

Assumes Cox PH model specification.

Details

All measures implemented from survAUC should be used with care, we are aware of problems in implementation that sometimes cause fatal errors in R. In future updates some of these measures may be re-written and implemented directly in mlr3proba.

Dictionary

This Measure can be instantiated via the dictionary mlr_measures or with the associated sugar function msr():

MeasureSurvXuR2$new()
mlr_measures$get("surv.xu_r2")
msr("surv.xu_r2")

Parameters

Empty ParamSet

Meta Information

Type: "surv"

Range: \([0, 1]\)

Minimize: FALSE

Required prediction: lp

References

Xu R, O'Quigley J (1999). “A R2 type measure of dependence for proportional hazards models.” Journal of Nonparametric Statistics, 12(1), 83--107. doi:10.1080/10485259908832799 .

See also

Super classes

mlr3::Measure -> mlr3proba::MeasureSurv -> MeasureSurvXuR2

Methods

Public methods

MeasureSurvXuR2$new()

MeasureSurvXuR2$clone()

Inherited methods

mlr3::Measure$aggregate()

mlr3::Measure$format()

mlr3::Measure$help()

mlr3::Measure$score()

mlr3proba::MeasureSurv$print()

Method new()

Creates a new instance of this R6 class.

Usage

MeasureSurvXuR2$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

MeasureSurvXuR2$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.