PipeOpPredClassifSurvDiscTime

Transform PredictionClassif to PredictionSurv by converting event probabilities of a pseudo status variable (discrete time hazards) to survival probabilities using the product rule (Tutz et al. 2016):

$$P_k = p_k\cdot ... \cdot p_1$$

Where:

We assume that continuous time is divided into time intervals \([0, t_1), [t_1, t_2), ..., [t_n, \infty)\)

\(P_k = P(T > t_k)\) is the survival probability at time \(t_k\)

\(h_k\) is the discrete-time hazard (classifier prediction), i.e. the conditional probability for an event in the \(k\)-interval.

\(p_k = 1 - h_k = P(T \ge t_k | T \ge t_{k-1})\)

Dictionary

This PipeOp can be instantiated via the dictionary mlr3pipelines::mlr_pipeops or with the associated sugar function mlr3pipelines::po():

PipeOpPredClassifSurvDiscTime$new()
mlr_pipeops$get("trafopred_classifsurv_disctime")
po("trafopred_classifsurv_disctime")

Input and Output Channels

The input is a PredictionClassif and a data.table with the transformed data both generated by PipeOpTaskSurvClassifDiscTime. The output is the input PredictionClassif transformed to a PredictionSurv. Only works during prediction phase.

References

Tutz, Gerhard, Schmid, Matthias (2016). Modeling Discrete Time-to-Event Data, series Springer Series in Statistics. Springer International Publishing. ISBN 978-3-319-28156-8 978-3-319-28158-2, http://link.springer.com/10.1007/978-3-319-28158-2.

See also

Super class

mlr3pipelines::PipeOp -> PipeOpPredClassifSurvDiscTime

Methods

Public methods

PipeOpPredClassifSurvDiscTime$new()

PipeOpPredClassifSurvDiscTime$clone()

Inherited methods

mlr3pipelines::PipeOp$help()

mlr3pipelines::PipeOp$predict()

mlr3pipelines::PipeOp$print()

mlr3pipelines::PipeOp$train()

Method new()

Creates a new instance of this R6 class.

Usage

PipeOpPredClassifSurvDiscTime$new(id = "trafopred_classifsurv_disctime")

Arguments

id

(character(1))
Identifier of the resulting object.

Method clone()

The objects of this class are cloneable with this method.

Usage

PipeOpPredClassifSurvDiscTime$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.