Density Task Generator for Package 'distr6'

A mlr3::TaskGenerator calling distr6::distrSimulate(). See distr6::distrSimulate() for an explanation of the hyperparameters. See distr6::listDistributions() for the names of the available distributions.

Dictionary

This TaskGenerator can be instantiated via the dictionary mlr_task_generators or with the associated sugar function tgen():

mlr_task_generators$get("simdens")
tgen("simdens")

Parameters

Id	Type	Default	Levels

distribution	character	Normal	Arcsine, Arrdist, Bernoulli, Beta, BetaNoncentral, Binomial, Categorical, Cauchy, ChiSquared, ChiSquaredNoncentral, ...

pars	untyped	-

See also

Super class

mlr3::TaskGenerator -> TaskGeneratorSimdens

Methods

Public methods

TaskGeneratorSimdens$new()

TaskGeneratorSimdens$help()

TaskGeneratorSimdens$clone()

Inherited methods

mlr3::TaskGenerator$format()

mlr3::TaskGenerator$generate()

mlr3::TaskGenerator$print()

Method new()

Creates a new instance of this R6 class.

Usage

TaskGeneratorSimdens$new()

Method help()

Opens the corresponding help page referenced by field $man.

Usage

TaskGeneratorSimdens$help()

Method clone()

The objects of this class are cloneable with this method.

Usage

TaskGeneratorSimdens$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# generate 20 samples from a standard Normal distribution
dens_gen = tgen("simdens")
dens_gen$param_set
#> <ParamSet(2)>
#>              id    class lower upper nlevels        default  value
#>          <char>   <char> <num> <num>   <num>         <list> <list>
#> 1: distribution ParamFct    NA    NA      47         Normal [NULL]
#> 2:         pars ParamUty    NA    NA     Inf <NoDefault[0]> [NULL]

task = dens_gen$generate(20)
head(task)
#>             x
#>         <num>
#> 1: -0.7994923
#> 2: -0.0871060
#> 3:  1.2420805
#> 4:  1.1361390
#> 5: -1.1946120
#> 6: -0.4110691

# generate 50 samples from a Binomial distribution with specific parameters
dens_gen = tgen("simdens", distribution = "Bernoulli", pars = list(prob = 0.8))
task = dens_gen$generate(50)
task$data()[["x"]]
#>  [1] 1 0 0 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 0 1 1 1 1
#> [39] 0 1 1 1 0 1 1 1 1 1 1 0