Get Survival Predict Types

Internal helper function to easily return the correct survival predict types.

Usage

surv_return(
  times = NULL,
  surv = NULL,
  crank = NULL,
  lp = NULL,
  response = NULL,
  which.curve = NULL
)

Arguments

times

(numeric())
Vector of survival times.

surv

(matrix()|array())
Matrix or array of predicted survival probabilities, rows (1st dimension) are observations, columns (2nd dimension) are times and in the case of an array there should be one more dimension. Number of columns should be equal to length of times. In case a numeric() vector is provided, it is converted to a single row (one observation) matrix.

crank

(numeric())
Relative risk/continuous ranking. Higher value is associated with higher risk. If NULL then either set as -response if available or lp if available (this assumes that the lp prediction comes from a PH type model - in case of an AFT model the user should provide -lp). In case neither response or lp are provided, then crank is calculated as the sum of the cumulative hazard function (expected mortality) derived from the predicted survival function (surv), see get_mortality. In case surv is a 3d array, we use the which.curve parameter to decide which survival matrix (index in the 3rd dimension) will be chosen for the calculation of crank.

lp

(numeric())
Predicted linear predictor, used to impute crank if NULL.

response

(numeric())
Predicted survival time, passed through function without modification.

which.curve

Which curve (3rd dimension) should the crank be calculated for, in case surv is an array? If between (0,1) it is taken as the quantile of the curves otherwise if greater than 1 it is taken as the curve index. It can also be 'mean' and the survival probabilities are averaged across the 3rd dimension. Default value (NULL) is the 0.5 quantile which is the median across the 3rd dimension of the survival array.

References

Sonabend, Raphael, Bender, Andreas, Vollmer, Sebastian (2022). “Avoiding C-hacking when evaluating survival distribution predictions with discrimination measures.” Bioinformatics. ISSN 1367-4803, doi:10.1093/BIOINFORMATICS/BTAC451 , https://academic.oup.com/bioinformatics/advance-article/doi/10.1093/bioinformatics/btac451/6640155.

Examples

n = 10 # number of observations
k = 50 # time points

# Create the matrix with random values between 0 and 1
mat = matrix(runif(n * k, min = 0, max = 1), nrow = n, ncol = k)

# transform it to a survival matrix
surv_mat = t(apply(mat, 1L, function(row) sort(row, decreasing = TRUE)))

# crank is expected mortality, distr is the survival matrix
surv_return(times = 1:k, surv = surv_mat)
#> $distr
#>               1         2         3         4         5         6         7
#>  [1,] 0.9794421 0.9589496 0.9476095 0.9402473 0.9382224 0.9288087 0.9178913
#>  [2,] 0.9934465 0.9805039 0.9485569 0.9414456 0.8775505 0.8738338 0.8336541
#>  [3,] 0.9851348 0.9451559 0.9386796 0.9241046 0.9134122 0.8984507 0.8790326
#>  [4,] 0.9713488 0.9498454 0.9421589 0.9334155 0.8897823 0.8839672 0.8838665
#>  [5,] 0.9973516 0.9588250 0.9200535 0.8911841 0.8785894 0.8688132 0.8364717
#>  [6,] 0.9552722 0.9502730 0.9465847 0.9291677 0.9254983 0.9119104 0.9025723
#>  [7,] 0.9603731 0.9177559 0.8983777 0.8971756 0.8917715 0.8748753 0.7848713
#>  [8,] 0.9869357 0.9664360 0.9656714 0.9508665 0.9244410 0.9018152 0.8951614
#>  [9,] 0.9948064 0.9728233 0.9502516 0.9201530 0.9094984 0.8864396 0.8687632
#> [10,] 0.9816401 0.9657578 0.9538755 0.9331108 0.9250900 0.9117160 0.9048754
#>               8         9        10        11        12        13        14
#>  [1,] 0.9149762 0.8926859 0.8918885 0.8576335 0.8321295 0.8114275 0.8079368
#>  [2,] 0.7906483 0.7899037 0.7814948 0.7574370 0.7559202 0.7518906 0.7393656
#>  [3,] 0.8753550 0.8718804 0.8625922 0.8324023 0.7986867 0.7857293 0.7550579
#>  [4,] 0.8291470 0.8247514 0.8136620 0.7983693 0.7879770 0.7530031 0.6559530
#>  [5,] 0.8263965 0.7616203 0.7535745 0.7455548 0.7450209 0.7432131 0.6930181
#>  [6,] 0.8973003 0.8971612 0.8748946 0.8737686 0.8512747 0.7645142 0.7271018
#>  [7,] 0.7510135 0.7461816 0.7450308 0.7431170 0.7399317 0.7204881 0.7023608
#>  [8,] 0.8929633 0.8836490 0.8357708 0.8245717 0.7912500 0.7689546 0.7657297
#>  [9,] 0.7988234 0.7859585 0.7603560 0.7555974 0.7339593 0.6805833 0.6739445
#> [10,] 0.8876921 0.8790991 0.8103576 0.8079030 0.7962478 0.7556817 0.7418578
#>              15        16        17        18        19        20        21
#>  [1,] 0.7943113 0.7872071 0.7837090 0.7371026 0.6396291 0.6060935 0.5856575
#>  [2,] 0.6795759 0.6758396 0.6738712 0.6542321 0.6444727 0.4797272 0.4653955
#>  [3,] 0.6826956 0.6663225 0.6551764 0.6510571 0.6224973 0.6178737 0.6112889
#>  [4,] 0.6074018 0.5759659 0.5707096 0.5437565 0.5349464 0.5275874 0.5099030
#>  [5,] 0.6928062 0.6784455 0.6466827 0.6449655 0.6109748 0.5878736 0.5651089
#>  [6,] 0.6871244 0.6837311 0.6820931 0.6771909 0.6650697 0.6508657 0.6352778
#>  [7,] 0.6739790 0.6459861 0.6094268 0.6084599 0.5848297 0.5783179 0.5691045
#>  [8,] 0.7408996 0.7267602 0.6921076 0.6777415 0.6771304 0.6493985 0.5742664
#>  [9,] 0.6561616 0.6455796 0.6148179 0.6128187 0.5083138 0.5032634 0.4879327
#> [10,] 0.7394767 0.6851048 0.6755378 0.6500370 0.6432805 0.6391245 0.6238111
#>              22        23        24        25        26        27        28
#>  [1,] 0.5697286 0.5626033 0.5583685 0.5367521 0.4990423 0.4936879 0.4890213
#>  [2,] 0.4551764 0.4515833 0.4077304 0.4014720 0.3853751 0.3780918 0.3693564
#>  [3,] 0.5606844 0.5395828 0.5383968 0.5316910 0.5198593 0.5087021 0.5086005
#>  [4,] 0.4806851 0.4687663 0.4561490 0.4489037 0.4372114 0.3869841 0.3840693
#>  [5,] 0.5203331 0.4987405 0.4793220 0.4719111 0.4560331 0.4493606 0.4013732
#>  [6,] 0.5682289 0.5396771 0.5191415 0.5190614 0.4957747 0.4936730 0.4674301
#>  [7,] 0.5165107 0.5037661 0.4984194 0.4630104 0.4512468 0.4491007 0.3976840
#>  [8,] 0.5679996 0.5455865 0.5323396 0.5290906 0.5248725 0.5005500 0.4764852
#>  [9,] 0.4852524 0.4414441 0.4321226 0.4186333 0.4101809 0.3965883 0.3862159
#> [10,] 0.5902491 0.5805523 0.5650550 0.5635035 0.5616572 0.5270831 0.5085274
#>              29        30        31        32        33        34        35
#>  [1,] 0.4419958 0.4268045 0.4260984 0.4154621 0.4010234 0.4003626 0.3981447
#>  [2,] 0.3580417 0.3541999 0.3036054 0.2820016 0.2764327 0.2546313 0.2449196
#>  [3,] 0.4962765 0.4895716 0.4852767 0.4787824 0.4427540 0.4288217 0.4136030
#>  [4,] 0.3730068 0.3684651 0.3611331 0.3560426 0.3175587 0.3063475 0.3026925
#>  [5,] 0.3956111 0.3794858 0.3725176 0.3689723 0.3537034 0.3412039 0.3051360
#>  [6,] 0.4655611 0.4612774 0.4588419 0.4452369 0.4422927 0.4412757 0.3224656
#>  [7,] 0.3881853 0.3499623 0.3168865 0.3033649 0.2638359 0.2586417 0.2477542
#>  [8,] 0.4628624 0.4492771 0.4452284 0.4314997 0.4181183 0.4166327 0.3761993
#>  [9,] 0.3720130 0.3576476 0.3426062 0.3178792 0.2802023 0.2643787 0.2598350
#> [10,] 0.4974766 0.4831529 0.4586042 0.4501553 0.4180359 0.4104929 0.3992938
#>              36        37        38        39        40        41        42
#>  [1,] 0.3395265 0.3067819 0.2848929 0.2775774 0.2754919 0.2725301 0.2491995
#>  [2,] 0.2397378 0.2346270 0.2344430 0.2297689 0.2296534 0.1933530 0.1697556
#>  [3,] 0.3782864 0.3529440 0.3105996 0.2956396 0.2816088 0.2693629 0.2061662
#>  [4,] 0.2553529 0.2454362 0.1948725 0.1811186 0.1790988 0.1490338 0.1324637
#>  [5,] 0.2920489 0.2264507 0.2125656 0.1850746 0.1531799 0.1468652 0.1392960
#>  [6,] 0.2976629 0.2878388 0.2856305 0.2825226 0.2708835 0.2452659 0.2144562
#>  [7,] 0.2370641 0.1997467 0.1878583 0.1416267 0.1320904 0.1218723 0.1155634
#>  [8,] 0.3599979 0.3206747 0.2580953 0.2525893 0.2485505 0.2117358 0.1701321
#>  [9,] 0.2526814 0.2190179 0.1503646 0.1299040 0.1145584 0.1071983 0.1038168
#> [10,] 0.3848675 0.3699081 0.3203076 0.2899586 0.2862637 0.2840219 0.2832142
#>               43         44         45         46         47         48
#>  [1,] 0.24519330 0.22377536 0.18309978 0.15387403 0.11552915 0.11384417
#>  [2,] 0.15390631 0.09976125 0.09040829 0.07244283 0.04275729 0.03423153
#>  [3,] 0.20321506 0.19073190 0.17899560 0.17539133 0.12397000 0.07337627
#>  [4,] 0.07509006 0.06878595 0.06172252 0.04879588 0.04642063 0.04095240
#>  [5,] 0.12945870 0.09388544 0.08024527 0.07321158 0.06928340 0.03836329
#>  [6,] 0.18563156 0.17848638 0.14342399 0.09629826 0.04491642 0.02176485
#>  [7,] 0.10790638 0.09659364 0.09388124 0.08266774 0.08179067 0.05090457
#>  [8,] 0.11687595 0.11187933 0.10463948 0.07157586 0.05980513 0.05698627
#>  [9,] 0.09637994 0.08755579 0.07884668 0.06546150 0.04937604 0.04564512
#> [10,] 0.25902884 0.25809975 0.24130251 0.22691089 0.16309018 0.13050936
#>               49          50
#>  [1,] 0.05280189 0.046426695
#>  [2,] 0.02960694 0.007672377
#>  [3,] 0.06784520 0.035663854
#>  [4,] 0.03257335 0.018953650
#>  [5,] 0.02948579 0.020708095
#>  [6,] 0.01666817 0.003846681
#>  [7,] 0.04215936 0.015561963
#>  [8,] 0.03487221 0.018326628
#>  [9,] 0.03628082 0.006419373
#> [10,] 0.11649351 0.074129862
#> 
#> $crank
#>  [1] 40.17930 55.69059 40.84215 56.34237 53.02445 48.60327 55.40242 46.88580
#>  [9] 58.74630 36.88353
#> 
#> $lp
#> NULL
#> 
#> $response
#> NULL
#> 

# if crank is set, it's not overwritten
surv_return(times = 1:k, surv = surv_mat, crank = rnorm(n))
#> $distr
#>               1         2         3         4         5         6         7
#>  [1,] 0.9794421 0.9589496 0.9476095 0.9402473 0.9382224 0.9288087 0.9178913
#>  [2,] 0.9934465 0.9805039 0.9485569 0.9414456 0.8775505 0.8738338 0.8336541
#>  [3,] 0.9851348 0.9451559 0.9386796 0.9241046 0.9134122 0.8984507 0.8790326
#>  [4,] 0.9713488 0.9498454 0.9421589 0.9334155 0.8897823 0.8839672 0.8838665
#>  [5,] 0.9973516 0.9588250 0.9200535 0.8911841 0.8785894 0.8688132 0.8364717
#>  [6,] 0.9552722 0.9502730 0.9465847 0.9291677 0.9254983 0.9119104 0.9025723
#>  [7,] 0.9603731 0.9177559 0.8983777 0.8971756 0.8917715 0.8748753 0.7848713
#>  [8,] 0.9869357 0.9664360 0.9656714 0.9508665 0.9244410 0.9018152 0.8951614
#>  [9,] 0.9948064 0.9728233 0.9502516 0.9201530 0.9094984 0.8864396 0.8687632
#> [10,] 0.9816401 0.9657578 0.9538755 0.9331108 0.9250900 0.9117160 0.9048754
#>               8         9        10        11        12        13        14
#>  [1,] 0.9149762 0.8926859 0.8918885 0.8576335 0.8321295 0.8114275 0.8079368
#>  [2,] 0.7906483 0.7899037 0.7814948 0.7574370 0.7559202 0.7518906 0.7393656
#>  [3,] 0.8753550 0.8718804 0.8625922 0.8324023 0.7986867 0.7857293 0.7550579
#>  [4,] 0.8291470 0.8247514 0.8136620 0.7983693 0.7879770 0.7530031 0.6559530
#>  [5,] 0.8263965 0.7616203 0.7535745 0.7455548 0.7450209 0.7432131 0.6930181
#>  [6,] 0.8973003 0.8971612 0.8748946 0.8737686 0.8512747 0.7645142 0.7271018
#>  [7,] 0.7510135 0.7461816 0.7450308 0.7431170 0.7399317 0.7204881 0.7023608
#>  [8,] 0.8929633 0.8836490 0.8357708 0.8245717 0.7912500 0.7689546 0.7657297
#>  [9,] 0.7988234 0.7859585 0.7603560 0.7555974 0.7339593 0.6805833 0.6739445
#> [10,] 0.8876921 0.8790991 0.8103576 0.8079030 0.7962478 0.7556817 0.7418578
#>              15        16        17        18        19        20        21
#>  [1,] 0.7943113 0.7872071 0.7837090 0.7371026 0.6396291 0.6060935 0.5856575
#>  [2,] 0.6795759 0.6758396 0.6738712 0.6542321 0.6444727 0.4797272 0.4653955
#>  [3,] 0.6826956 0.6663225 0.6551764 0.6510571 0.6224973 0.6178737 0.6112889
#>  [4,] 0.6074018 0.5759659 0.5707096 0.5437565 0.5349464 0.5275874 0.5099030
#>  [5,] 0.6928062 0.6784455 0.6466827 0.6449655 0.6109748 0.5878736 0.5651089
#>  [6,] 0.6871244 0.6837311 0.6820931 0.6771909 0.6650697 0.6508657 0.6352778
#>  [7,] 0.6739790 0.6459861 0.6094268 0.6084599 0.5848297 0.5783179 0.5691045
#>  [8,] 0.7408996 0.7267602 0.6921076 0.6777415 0.6771304 0.6493985 0.5742664
#>  [9,] 0.6561616 0.6455796 0.6148179 0.6128187 0.5083138 0.5032634 0.4879327
#> [10,] 0.7394767 0.6851048 0.6755378 0.6500370 0.6432805 0.6391245 0.6238111
#>              22        23        24        25        26        27        28
#>  [1,] 0.5697286 0.5626033 0.5583685 0.5367521 0.4990423 0.4936879 0.4890213
#>  [2,] 0.4551764 0.4515833 0.4077304 0.4014720 0.3853751 0.3780918 0.3693564
#>  [3,] 0.5606844 0.5395828 0.5383968 0.5316910 0.5198593 0.5087021 0.5086005
#>  [4,] 0.4806851 0.4687663 0.4561490 0.4489037 0.4372114 0.3869841 0.3840693
#>  [5,] 0.5203331 0.4987405 0.4793220 0.4719111 0.4560331 0.4493606 0.4013732
#>  [6,] 0.5682289 0.5396771 0.5191415 0.5190614 0.4957747 0.4936730 0.4674301
#>  [7,] 0.5165107 0.5037661 0.4984194 0.4630104 0.4512468 0.4491007 0.3976840
#>  [8,] 0.5679996 0.5455865 0.5323396 0.5290906 0.5248725 0.5005500 0.4764852
#>  [9,] 0.4852524 0.4414441 0.4321226 0.4186333 0.4101809 0.3965883 0.3862159
#> [10,] 0.5902491 0.5805523 0.5650550 0.5635035 0.5616572 0.5270831 0.5085274
#>              29        30        31        32        33        34        35
#>  [1,] 0.4419958 0.4268045 0.4260984 0.4154621 0.4010234 0.4003626 0.3981447
#>  [2,] 0.3580417 0.3541999 0.3036054 0.2820016 0.2764327 0.2546313 0.2449196
#>  [3,] 0.4962765 0.4895716 0.4852767 0.4787824 0.4427540 0.4288217 0.4136030
#>  [4,] 0.3730068 0.3684651 0.3611331 0.3560426 0.3175587 0.3063475 0.3026925
#>  [5,] 0.3956111 0.3794858 0.3725176 0.3689723 0.3537034 0.3412039 0.3051360
#>  [6,] 0.4655611 0.4612774 0.4588419 0.4452369 0.4422927 0.4412757 0.3224656
#>  [7,] 0.3881853 0.3499623 0.3168865 0.3033649 0.2638359 0.2586417 0.2477542
#>  [8,] 0.4628624 0.4492771 0.4452284 0.4314997 0.4181183 0.4166327 0.3761993
#>  [9,] 0.3720130 0.3576476 0.3426062 0.3178792 0.2802023 0.2643787 0.2598350
#> [10,] 0.4974766 0.4831529 0.4586042 0.4501553 0.4180359 0.4104929 0.3992938
#>              36        37        38        39        40        41        42
#>  [1,] 0.3395265 0.3067819 0.2848929 0.2775774 0.2754919 0.2725301 0.2491995
#>  [2,] 0.2397378 0.2346270 0.2344430 0.2297689 0.2296534 0.1933530 0.1697556
#>  [3,] 0.3782864 0.3529440 0.3105996 0.2956396 0.2816088 0.2693629 0.2061662
#>  [4,] 0.2553529 0.2454362 0.1948725 0.1811186 0.1790988 0.1490338 0.1324637
#>  [5,] 0.2920489 0.2264507 0.2125656 0.1850746 0.1531799 0.1468652 0.1392960
#>  [6,] 0.2976629 0.2878388 0.2856305 0.2825226 0.2708835 0.2452659 0.2144562
#>  [7,] 0.2370641 0.1997467 0.1878583 0.1416267 0.1320904 0.1218723 0.1155634
#>  [8,] 0.3599979 0.3206747 0.2580953 0.2525893 0.2485505 0.2117358 0.1701321
#>  [9,] 0.2526814 0.2190179 0.1503646 0.1299040 0.1145584 0.1071983 0.1038168
#> [10,] 0.3848675 0.3699081 0.3203076 0.2899586 0.2862637 0.2840219 0.2832142
#>               43         44         45         46         47         48
#>  [1,] 0.24519330 0.22377536 0.18309978 0.15387403 0.11552915 0.11384417
#>  [2,] 0.15390631 0.09976125 0.09040829 0.07244283 0.04275729 0.03423153
#>  [3,] 0.20321506 0.19073190 0.17899560 0.17539133 0.12397000 0.07337627
#>  [4,] 0.07509006 0.06878595 0.06172252 0.04879588 0.04642063 0.04095240
#>  [5,] 0.12945870 0.09388544 0.08024527 0.07321158 0.06928340 0.03836329
#>  [6,] 0.18563156 0.17848638 0.14342399 0.09629826 0.04491642 0.02176485
#>  [7,] 0.10790638 0.09659364 0.09388124 0.08266774 0.08179067 0.05090457
#>  [8,] 0.11687595 0.11187933 0.10463948 0.07157586 0.05980513 0.05698627
#>  [9,] 0.09637994 0.08755579 0.07884668 0.06546150 0.04937604 0.04564512
#> [10,] 0.25902884 0.25809975 0.24130251 0.22691089 0.16309018 0.13050936
#>               49          50
#>  [1,] 0.05280189 0.046426695
#>  [2,] 0.02960694 0.007672377
#>  [3,] 0.06784520 0.035663854
#>  [4,] 0.03257335 0.018953650
#>  [5,] 0.02948579 0.020708095
#>  [6,] 0.01666817 0.003846681
#>  [7,] 0.04215936 0.015561963
#>  [8,] 0.03487221 0.018326628
#>  [9,] 0.03628082 0.006419373
#> [10,] 0.11649351 0.074129862
#> 
#> $crank
#>  [1]  1.5917256 -0.6396042  0.2472043  0.0407682 -0.2606190  0.4690159
#>  [7] -0.3393448 -0.4227681  0.5258209  1.0690263
#> 
#> $lp
#> NULL
#> 
#> $response
#> NULL
#> 

# lp = crank
surv_return(lp = rnorm(n))
#> $distr
#> NULL
#> 
#> $crank
#>  [1]  0.83283003  0.01460201 -0.55357832 -1.23916072  1.23083682  1.07198008
#>  [7]  0.31167948  1.23201534  0.69029434  0.12754167
#> 
#> $lp
#>  [1]  0.83283003  0.01460201 -0.55357832 -1.23916072  1.23083682  1.07198008
#>  [7]  0.31167948  1.23201534  0.69029434  0.12754167
#> 
#> $response
#> NULL
#> 

# if response is set and no crank, crank = -response
surv_return(response = sample(1:100, n))
#> $distr
#> NULL
#> 
#> $crank
#>  [1]  -8 -57 -31 -87 -39  -5 -97 -58 -56 -10
#> 
#> $lp
#> NULL
#> 
#> $response
#>  [1]  8 57 31 87 39  5 97 58 56 10
#> 

# if both are set, they are not overwritten
surv_return(crank = rnorm(n), response = sample(1:100, n))
#> $distr
#> NULL
#> 
#> $crank
#>  [1]  0.191266878  0.138724083 -0.232420256 -1.454757541 -0.168093406
#>  [6] -0.096943595 -0.834856562  1.097533433 -0.739561382 -0.004608769
#> 
#> $lp
#> NULL
#> 
#> $response
#>  [1]  27   1  42   6  29 100  16  84  65  37
#>