Abstract Class for survAUC Measures

This is an abstract class that should not be constructed directly.

Parameter details

integrated (logical(1))
If TRUE (default), returns the integrated score (eg across time points); otherwise, not integrated (eg at a single time point).

times (numeric())
If integrate == TRUE then a vector of time-points over which to integrate the score. If integrate == FALSE then a single time point at which to return the score.

Super classes

mlr3::Measure -> mlr3proba::MeasureSurv -> MeasureSurvAUC

Methods

Public methods

MeasureSurvAUC$new()

MeasureSurvAUC$clone()

Inherited methods

mlr3::Measure$aggregate()

mlr3::Measure$format()

mlr3::Measure$help()

mlr3::Measure$score()

mlr3proba::MeasureSurv$print()

Method new()

Creates a new instance of this R6 class.

Usage

MeasureSurvAUC$new(
  id,
  properties = character(),
  label = NA_character_,
  man = NA_character_,
  param_set = ps()
)

Arguments

id

(character(1))
Identifier for the new instance.

properties

(character())
Properties of the measure. Must be a subset of mlr_reflections$measure_properties. Supported by mlr3:

"requires_task" (requires the complete Task),

"requires_learner" (requires the trained Learner),

"requires_train_set" (requires the training indices from the Resampling), and

"na_score" (the measure is expected to occasionally return NA or NaN).

label

(character(1))
Label for the new instance.

man

(character(1))
String in the format [pkg]::[topic] pointing to a manual page for this object. The referenced help package can be opened via method $help().

param_set

(paradox::ParamSet)
Set of hyperparameters.

Method clone()

The objects of this class are cloneable with this method.

Usage

MeasureSurvAUC$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.