Density Task

This task specializes TaskUnsupervised for density estimation problems. The data in backend should be a numeric vector or a one column matrix-like object. The task_type is set to "density".

Predefined tasks are stored in the dictionary mlr_tasks.

See also

Super classes

mlr3::Task -> mlr3::TaskUnsupervised -> TaskDens

Methods

Public methods

TaskDens$new()

TaskDens$clone()

Inherited methods

mlr3::Task$add_strata()

mlr3::Task$cbind()

mlr3::Task$data()

mlr3::Task$droplevels()

mlr3::Task$filter()

mlr3::Task$format()

mlr3::Task$formula()

mlr3::Task$head()

mlr3::Task$help()

mlr3::Task$levels()

mlr3::Task$missings()

mlr3::Task$print()

mlr3::Task$rbind()

mlr3::Task$rename()

mlr3::Task$select()

mlr3::Task$set_col_roles()

mlr3::Task$set_row_roles()

Method new()

Creates a new instance of this R6 class.

Usage

TaskDens$new(id, backend)

Arguments

id

(character(1))
Identifier for the new instance.

backend

(DataBackend)
Either a DataBackend, a matrix-like object, or a numeric vector. If weights are used then two columns expected, otherwise one column. The weight column must be clearly specified (via [Task]$col_roles) or the learners will break.

Method clone()

The objects of this class are cloneable with this method.

Usage

TaskDens$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

task = TaskDens$new("precip", backend = precip)
task$task_type
#> [1] "dens"