Log loss Density Measure

Calculates the cross-entropy, or logarithmic (log), loss.

The logloss, in the context of probabilistic predictions, is defined as the negative log probability density function, \(f\), evaluated at the observed value, \(y\), $$L(f, y) = -\log(f(y))$$

Meta Information

Type: "density"

Range: \([0, \infty)\)

Minimize: TRUE

Required prediction: pdf

Super classes

mlr3::Measure -> mlr3proba::MeasureDens -> MeasureDensLogloss

Methods

Public methods

MeasureDensLogloss$new()

MeasureDensLogloss$clone()

Inherited methods

mlr3::Measure$aggregate()

mlr3::Measure$format()

mlr3::Measure$help()

mlr3::Measure$print()

mlr3::Measure$score()

Method new()

Creates a new instance of this R6 class.

Usage

MeasureDensLogloss$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

MeasureDensLogloss$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.