Survival Task Generator for Package 'simsurv'

A mlr3::TaskGenerator calling simsurv::simsurv() from package simsurv.

This generator currently only exposes a small subset of the flexibility of simsurv, and just creates a small data set with the following numerical covariates:

treatment: Bernoulli distributed with log hazard ratio -0.5.

height: Normally distributed with log hazard ratio 1.

weight: normally distributed with log hazard ratio 0.

See simsurv::simsurv() for an explanation of the hyperparameters.

Dictionary

This TaskGenerator can be instantiated via the dictionary mlr_task_generators or with the associated sugar function tgen():

mlr_task_generators$get("simsurv")
tgen("simsurv")

See also

Super class

mlr3::TaskGenerator -> TaskGeneratorSimsurv

Methods

Public methods

TaskGeneratorSimsurv$new()

TaskGeneratorSimsurv$clone()

Inherited methods

mlr3::TaskGenerator$format()

mlr3::TaskGenerator$generate()

mlr3::TaskGenerator$print()

Method new()

Creates a new instance of this R6 class.

Usage

TaskGeneratorSimsurv$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

TaskGeneratorSimsurv$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("simsurv", quietly = TRUE)) {
  generator = mlr3::mlr_task_generators$get("simsurv")
  task = generator$generate(20)
  task$head()
}
#>    eventtime status   height treatment   weight
#> 1:  2.393268      1 194.0304         1 88.62086
#> 2:  1.243874      1 182.6473         0 77.56763
#> 3:  5.000000      0 183.6553         1 77.93913
#> 4:  5.000000      0 204.3532         1 80.19178
#> 5:  5.000000      0 181.6806         0 80.29561
#> 6:  2.490183      1 177.9900         1 85.49828